Back off when server fails (HTTP 50x)
When the GitHub API replies with a server error (usually HTTP 502), retry up to `maxServerErrors` times with `serverErrorBackOff` intervals in between.

Currently, the inputs above are constants: each request will tolerate at most 10 errors and will wait 10 seconds between each attempt (this is _in addition_ to any rate limiting imposed by the GitHub API). If needed, it is easy enough to add both as optional arguments to `fetchStreamed`.

This is important for long-running batch jobs. Without this change, the package will just crash any time it encounteres a GitHub HTTP 50x error (which are rare, but not unheard of), losing any pagination progress.